Fix failures and inline shared.ts
Remove fake-promise types

WIP
