+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", "dist", "types", "test"],
+  "exclude": ["node_modules", "dist", "test"],
   "compilerOptions": {
     "target": "ES2020",
     "module": "ESNext",
@@ -9,7 +9,6 @@
     "rootDir": "src",
 
     "declaration": true,
-    "declarationMap": true,
     "outDir": "dist",
 
     "strict": true,
